requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
langchain==0.1.9
langchain-community==0.0.24
anthropic==0.19.1
requests==2.31.0
psycopg2-binary==2.9.9
